+++ Dataset/utils/dem_utils.py
@@ -1,11 +1,12 @@
 from demoparser2 import DemoParser #https://github.com/LaihoE/demoparser
+from csgo.sharecode import decode
 import pandas
 import json
 import time
 
 import pandas.core
 import pandas.core.frame
-from demo_parser_fields import ALL_FIELDS
+from utils.demo_parser_fields import ALL_FIELDS
 
 START_BALANCE = 800
 MAX_HEALTH = 100.0
@@ -76,54 +77,6 @@ def df_to_readable_csv(df:pandas.core.frame.DataFrame, filename:str):
         # Write the data rows (with commas and aligned)
         for index, row in df_aligned.iterrows():
             f.write(','.join([str(row[col]).ljust(max_lengths[col]) for col in df.columns]) + '\n')
-
-def binary_search_sorted_dataframe(df:pandas.core.frame.DataFrame, col, val)-> int:
-    """
-        Uses binary search to find the first occurance of val in a sorted column(col) of a dataframe(df)
-    """
-    def binary_search_helper(df:pandas.core.frame.DataFrame , low:int, high:int, col, val)-> int:
-        if high >= low:
-            mid = low + (high - low) // 2
-
-            if df[col].iloc[mid] == val:
-                # Check if this is the first occurrence
-                if mid == 0 or df[col].iloc[mid-1] != val:
-                    return mid
-                else:
-                    # Look to the left
-                    return binary_search_helper(df, low, mid-1, col, val)
-                
-            elif df[col].iloc[mid] > val:
-                # Look to the left
-                return binary_search_helper(df, low, mid-1, col, val)
-            else:
-                # Look to the right
-                return binary_search_helper(df, mid+1, high, col, val)
-        else:
-            return -1  # If no match is found
-        
-    return binary_search_helper(df, 0, len(df.index), col, val)
-
-def remove_warmup_rounds(df: pandas.core.frame.DataFrame) -> pandas.core.frame.DataFrame:
-    """removes the ticks corresponding to a warmup round for a given data frame"""
-
-    index = binary_search_sorted_dataframe(df,'is_warmup_period', False)
-    return df.loc[index:].reset_index(drop=True)
-
-def get_tick(df:pandas.core.frame.DataFrame, index:int) -> pandas.core.frame.DataFrame:
-    """Returns a single tick of the index from the dataframe"""
-
-    i = binary_search_sorted_dataframe(df, 'tick', index)
-
-    return df.loc[i:i+NUM_PLAYERS-1].reset_index(drop=True)
-
-def get_ticks(df:pandas.core.frame.DataFrame, start:int, end:int) -> pandas.core.frame.DataFrame:
-    """Returns a dataframe of ticks of the index from the dataframe"""
-
-    s = binary_search_sorted_dataframe(df, 'tick', start)
-    t = binary_search_sorted_dataframe(df, 'tick', end)
-
-    return df.loc[s:t+NUM_PLAYERS-1].reset_index(drop=True)
 
 def events_2_json(filename:str):
     """
@@ -220,6 +173,8 @@ def remove_sensitive_data_df(df:pandas.core.frame.DataFrame):
 
     return df_anonymised
 
-def create_cheater_df():
+def steamlink_to_sharecode(url:str):
+    return url.split("%20")[1]
 
-    df = { "name" : [] }
+def sharecode_to_outcomeid(share_code):
+    return decode(share_code)["outcomeid"]
